Refactor for maintainability into focused modules and use tinycolor
