+++ addons/paint-gradient-editor/addon.json
@@ -31,7 +31,7 @@
   "tags": ["editor", "costumeEditor"],
   "versionAdded": "1.45.0",
   "enabledByDefault": false,
-  "libraries": ["scratch-paint"],
+  "libraries": ["scratch-paint", "tinycolor2"],
   "userscripts": [
     {
       "url": "userscript.js",
